Initial commit
Committing this code exactly how it was Sunday evening, no changes, no comments, no commented out code removed.
The entire codebase for my submission for a game jam: Ryan Laley's 2023 Game Jam. The theme was connection, so I used the idea that making connections killed the enemies.
The main things I'd change with this are the enemies: they get their "MyEnemy" (CombatTarget) by getting PlayerCharacter 0, and that's obviously very static. I would make them a little "smarter" in that they currently fire at the PC if the PC is in range, regardless of whether the PC is behind a wall or if the PC is on a different "level" (level in the gameplay sense, not the UE maps sense).
Another thing is that the particle effects I have are limited and I'd have loved the projectiles to explode. The final thing is that there are no sounds or music at all and it's definitely to the game's detriment.
The main things I'm happy about are that I successfully figured out a connection system where ConnectionLines and ConnectionBoxes are in a singly-linked list, basically, and it's kind of cool. They both have a common parent, which is how it works. I tried using interfaces but my Googling said that UE's interfaces are not quite up to the job. It's unfortunate how complicated things get when dealing with c++ AND blueprints.