Fixed user not loading Main Menu after dying
So enemies would just fire at the character whether the character was alive or dead, which would cause the DeathTimer to continually be reset, causing the user to never be transported back to the main menu.
Enemies will no longer fire if the character is dead, and the character likewise will not take damage if they are dead.