Added Lives to HUD
Updated damage amounts to be units of 1 so it's more like lives.
Added a WBP Progress Bar and added copies to the screen; it has drain and fill animations and Drain() and Fill() functions. The Overlay keeps track of how many Progress Bars (lives) there are and drains them appropriately.
Health (Lives) is set by default to 4 and can be changed in the Character's BP.