Changed UI to be updated via delegates
Added a new GameInstance that manages the LevelManager and the HUD.
Added a new HUD that manages the Overlay and OverlayController
Added a new OverlayWidgetController that binds to callbacks by various game objects and rebroadcasts them. The Overlay has an OWC and binds to its callbacks, then runs the appropriate functions in response.
The Character does not have an Overlay object anymore and uses delegates to get information to the UI. The LevelManager now manages all of the level state and also uses delegates to get information to the UI.