Added particle effects
Replaced the projectile mesh with a particle effect.
Added a particle effect explosion.
Added "pollen" particle effects
Added a particle effect when the player connects the boxes.
Fixed a rotation bug when the player is transported to a new level.
Fixed connection boxes changing color even when they aren't 'ready'.
Added overlay text when the player gets near a connection box telling them how to use it.
Added flora to the levels and updated the lighting so they are not too dark.