Added loading screen and enemy firing options
Added a loading screen between the main menu and the main level.
Added ability for users to leave the level and return to the main menu using Esc.
Added different projectile types and made different orcs fire different types. As part of this, I composited an animation so the orc fires two projectiles in the same animation if they are set to fire sequentially.