Added a hit counter and removed player death
Changed the game so instead of the player having health and dying and losing when they run out of health, the player has a hit counter which increments when the player is hit by a projectile. This is the beginning of not allowing the player to move on to future levels if the hit count is too high.