Updated UI information about items
Updated the overlay text when the player comes near a connection box. If the player has all of the items the regular "Repair Defenses" text appears, and if the player does not then "Not enough items to repair!" text appears.
Updated the overlay to show the total count of items that the player needs to collect to make it clear to the player how many items of each type they need to find.