End Level when Player is hit too many times
Un-implemented player cannot die.
Turns out it doesn't make much sense to have the player complete the level and fix all the fence breaks and then be told that they didn't do it well enough and they have to do it again. So this adds functionality back in for the player to be pulled out of the level if they are hit too many times. But this time instead of dying, their supervisor pulls them out.